back to simple network stats
